btc_predict_nestquant-main/requirements.txt
pandas>=1.5.0
numpy>=1.24.0
lightgbm>=3.3.0
scikit-learn>=1.2.0
python-dotenv>=1.0.0
requests>=2.28.0
pytz>=2023.3
optuna>=3.0.0
ipython>=8.0.0

